Software_Testing_Project-Automation_Testing
Automation testing is beyond the scope. Functional testing & external interfaces are in scope and need to be tested. The scope of this project is limited to the testing of the features described in the succeeding sections of this document. Non-functional testing like stress, performance is beyond scope of this project.